+++ src/textField/index.unit.tsx
@@ -235,3 +235,16 @@ it('Simulates a blur event.', () => {
   wrapper.find('input').simulate('blur')
   expect(onTextFieldBlur).toHaveBeenCalledTimes(1)
 })
+
+it('Should format the default value', () => {
+  const wrapper = mount(
+    <TextField name="test" defaultValue="Hello" format={value => `${value} world`} />,
+  )
+  expect(wrapper.find('input').prop('value')).toBe('Hello world')
+})
+
+it('Should format the values when it changes', () => {
+  const wrapper = mount(<TextField name="test" format={value => `${value} world`} />)
+  wrapper.setState({ value: 'Hello' })
+  expect(wrapper.find('input').prop('value')).toBe('Hello world')
+})