+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 # v0.20.4 (10/01/2019)
 
 - **[FIX]** `SuccessModal` display with small content
+- **[UPDATE]** Allow `format` prop on TextField component
 
 # v0.20.3 (09/01/2019)